ci: run develop workflow on pushes to main
We want all pushes to main to be tested, but only those that change
relevant source code should trigger the release workflow. So, for pushes
to main where the release workflow isn't triggered, this one will help.
For the ones where the release workflow _is_ triggered it'll still run,
which creates some redundancy — I'll have to see if I can fix that
easily or not.