feat!(postcss-egal): revamp transformer
Revamp the transformer, fixing some bugs and improving warnings.

BREAKING CHANGE: Custom parsers are no longer supported. This is to
simplify the API and code a bit, and because the transformer supports ad
hoc options by now.

feat!(postcss-egal): revamp transformer

Revamp the transformer, fixing some bugs and improving warnings.

BREAKING CHANGE: Custom parsers are no longer supported. This is to
simplify the API and code a bit, and because the transformer supports ad
hoc options by now.